Revert our handler into a subclass.
